+++ src/app/forms/time-picker/time-picker.component.ts
@@ -19,7 +19,7 @@ import {
   animate
 } from '@angular/animations';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { timer } from 'rxjs/observable/timer';
+import { timer } from 'rxjs';
 
 export class FrTimePickerChange {
   source: FrTimePickerComponent;
@@ -47,6 +47,7 @@ export const TIME_PICKER_CONTROL_VALUE_ACCESSOR: any = {
 @Component({
   selector: 'fr-time-picker',
   templateUrl: './time-picker.component.html',
+  styleUrls: ['./time-picker.component.scss'],
   providers: [TIME_PICKER_CONTROL_VALUE_ACCESSOR],
   animations: [
     trigger('clockVisibility', [
@@ -74,13 +75,14 @@ export class FrTimePickerComponent implements OnInit, AfterViewInit, ControlValu
   @Output() change: EventEmitter<FrTimePickerChange> = new EventEmitter<FrTimePickerChange>();
 
   @ViewChild('clock') clock: ElementRef;
+  @ViewChild('clockHand') clockHand: ElementRef;
 
   @HostBinding('class.fr-timepicker-host') true;
 
   /**
    * For ControlValueAccessor
    */
-  private _innerValue: any;
+  private _innerValue: Date = new Date();
   private _onChangeCallback: (_: any) => void = noop;
   private _onTouchedCallback: () => void = noop;
   private _isDisabled = false;
@@ -98,14 +100,14 @@ export class FrTimePickerComponent implements OnInit, AfterViewInit, ControlValu
   }
 
   set value(obj: any) {
-    if (obj !== this._innerValue) {
+    if (obj !== this._innerValue && obj !== null && obj !== undefined) {
       this._innerValue = obj;
       this._onChangeCallback(obj);
     }
   }
 
   writeValue(obj: any): void {
-    if (obj !== this._innerValue) {
+    if (obj !== this._innerValue && obj !== null && obj !== undefined) {
       this._innerValue = obj;
     }
   }
@@ -169,14 +171,13 @@ export class FrTimePickerComponent implements OnInit, AfterViewInit, ControlValu
   }
 
   public putHandRightPosition(dial: number): void {
-    const hand = document.getElementById('clock-hand');
-    let deg  = 0;
+    let deg = 0;
     if (this.pickTarget === HOURS) {
       deg = (dial % 12) * 30;
     } else {
       deg = ((dial / 5) % 12) * 30;
     }
-    hand.style.transform = 'rotate(' + deg + 'deg)';
+    this.clockHand.nativeElement.style.transform = `rotate(${deg}deg)`;
   }
 
   public toggleTimePickerVisibility(): void {