+++ src/app/francette.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { RouterModule, Routes } from '@angular/router';
 
 // Francette modules
 import { FrButtonModule } from './button/button.module';
@@ -19,13 +18,13 @@ import { FrRippleModule } from './ripple/ripple.module';
 import { FrSideNavModule } from './side-nav/side-nav.module';
 import { FrTabsModule } from './tabs/tabs.module';
 import { FrToasterModule } from './toaster/toaster.module';
+import { FrTooltipModule } from './tooltip/tooltip.module';
 
 @NgModule({
   imports: [
     BrowserAnimationsModule,
     CommonModule,
     FormsModule,
-    RouterModule,
 
     FrButtonModule,
     FrChipModule,
@@ -39,7 +38,8 @@ import { FrToasterModule } from './toaster/toaster.module';
     FrRippleModule,
     FrSideNavModule,
     FrTabsModule,
-    FrToasterModule
+    FrToasterModule,
+    FrTooltipModule
   ],
   exports: [
     FrButtonModule,
@@ -54,7 +54,8 @@ import { FrToasterModule } from './toaster/toaster.module';
     FrRippleModule,
     FrSideNavModule,
     FrTabsModule,
-    FrToasterModule
+    FrToasterModule,
+    FrTooltipModule
   ]
 })
 export class FrancetteModule { }