+++ src/app/notification/notification.service.ts
@@ -10,12 +10,12 @@ import {
   OnInit,
   OnChanges,
   AfterViewInit,
-  AnimationTransitionEvent,
   HostListener,
   ReflectiveInjector,
   ComponentRef
 } from '@angular/core';
 import {
+  AnimationEvent,
   animate,
   trigger,
   style,
@@ -23,9 +23,7 @@ import {
   transition
 } from '@angular/animations';
 import { FrNotificationType, FrNotificationParam } from './notification.types';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
-import { timer } from 'rxjs/observable/timer';
+import { Observable, Observer, timer } from 'rxjs';
 
 export class FrNotificationContext<T> implements Observer<T> {
   constructor(
@@ -258,7 +256,7 @@ export class FrNotificationContentComponent implements OnInit {
     this.notificationState = 'void';
   }
 
-  public dismiss(event: AnimationTransitionEvent): void {
+  public dismiss(event: AnimationEvent): void {
     if (event.fromState === 'active' && event.toState === 'void') {
       if (!this.closed) {
         this._context.complete();