+++ src/app/progress/bar/bar.component.ts
@@ -1,22 +1,20 @@
-import { Component, OnInit, HostBinding } from '@angular/core';
+import { Component, Input, HostBinding } from '@angular/core';
 
 @Component({
   selector: 'fr-progress-bar',
   template: `
     <div class="fr-progress-bar">
       <div class="fr-progress-bar__background"></div>
-      <div class="fr-progress-bar__first-bar"></div>
-      <div class="fr-progress-bar__second-bar"></div>
+      <div class="fr-progress-bar__first-bar" [ngStyle]="{'background-color': color}"></div>
+      <div class="fr-progress-bar__second-bar" [ngStyle]="{'background-color': color}"></div>
     </div>
-  `
+  `,
+  styleUrls: ['./bar.component.scss']
 })
-export class FrProgressBarComponent implements OnInit {
+export class FrProgressBarComponent {
 
   @HostBinding('class.fr-progress-bar-host') true;
 
-  constructor() { }
-
-  ngOnInit() {
-  }
+  @Input() color = '#4E4F97';
 
 }