+++ src/app/toaster/toaster.service.ts
@@ -9,6 +9,7 @@ import {
   forwardRef,
   OnInit,
   OnChanges,
+  OnDestroy,
   AfterViewInit,
   HostListener,
   ReflectiveInjector,
@@ -21,10 +22,8 @@ import {
   transition,
   trigger
 } from '@angular/animations';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
-import { timer } from 'rxjs/observable/timer';
-
+import { Observable, Observer, timer } from 'rxjs';
+import { tap, delay, map } from 'rxjs/operators';
 import { FrToasterParam } from './toaster.types';
 
 export class FrToasterContext<T> implements Observer<T> {
@@ -76,7 +75,34 @@ export class FrToasterService {
     this.vcr = vcr;
   }
 
+  public pop<T>(toasterParam: FrToasterParam) {
+    const component = FrToasterContentComponent;
+    const componentFactory = this.cfr.resolveComponentFactory(component);
+
+    const noop = () => {};
+    const bindings = ReflectiveInjector.resolve([
+      { provide: FrToasterContext, useValue: new FrToasterContext(noop, noop, noop, toasterParam) }
+    ]);
+    const contextInjector = this.vcr.parentInjector;
+    const injector = ReflectiveInjector.fromResolvedProviders(bindings, contextInjector);
+
+    const componentRef = this.vcr.createComponent(componentFactory, this.vcr.length, injector);
+    this.vcr.element.nativeElement.appendChild(componentRef.location.nativeElement);
+    this.count++;
+
+    // Destroy toaster component with animation
+    timer((toasterParam.timeout || 3000) - 300).pipe(
+      tap(_ => componentRef.instance.toasterState = 'void'),
+      delay(300)  // This is for animation
+    ).subscribe(() => {
+      componentRef.destroy();
+    });
+
+    return componentRef.instance;
+  }
+
   public open<T>(toasterParam: FrToasterParam): Observable<T> {
+    console.warn('FrToasterService.open is DEPRECATED. Use FrToasterService.pop instead.');
     return new Observable<T>((observer: Observer<T>) => {
       const component = FrToasterContentComponent;
       const componentFactory = this.cfr.resolveComponentFactory(component);
@@ -84,19 +110,15 @@ export class FrToasterService {
       const _onNext = (value: T) => {
         if (componentRef) {
           if (observer.next) { observer.next(value); }
-          if (observer.complete) { observer.complete(); }
           observer.closed = true;
           this.count--;
-          componentRef.destroy();
         }
       };
       const _onError = (reason?: any) => {
         if (componentRef) {
           if (observer.error) { observer.error(reason); }
-          if (observer.complete) { observer.complete(); }
           observer.closed = true;
           this.count--;
-          componentRef.destroy();
         }
       };
       const _onComplete = (): void => {
@@ -104,7 +126,6 @@ export class FrToasterService {
           if (observer.complete) { observer.complete(); }
           observer.closed = true;
           this.count--;
-          componentRef.destroy();
         }
       };
       const bindings = ReflectiveInjector.resolve([
@@ -116,6 +137,14 @@ export class FrToasterService {
       const componentRef = this.vcr.createComponent(componentFactory, this.vcr.length, injector);
       this.vcr.element.nativeElement.appendChild(componentRef.location.nativeElement);
       this.count++;
+
+      // Destroy component after `toasterParam.timeout` milliseconds
+      timer((toasterParam.timeout || 3000) - 300).pipe(
+        tap(_ => componentRef.instance.toasterState = 'void'),
+        delay(300)    // This is for animation
+      ).subscribe(() => {
+        componentRef.destroy();
+      });
     });
   }
 }
@@ -208,13 +237,14 @@ export class FrToasterEntryComponent implements AfterViewInit {
     ])
   ]
 })
-export class FrToasterContentComponent implements OnInit {
+export class FrToasterContentComponent<T> implements OnInit, OnDestroy {
   public text = '';
   public action = '';
   public timeout = 500;
   public closed = false;
   public toasterState = 'void';
-  constructor (@Inject(forwardRef(() => FrToasterContext)) private _context: FrToasterContext<string|number>) {
+  public onActionObserver: Observer<any>;
+  constructor (@Inject(forwardRef(() => FrToasterContext)) private _context: FrToasterContext<T>) {
     this.text    = this._context.text;
     this.action  = this._context.action;
     this.timeout = this._context.timeout;
@@ -222,23 +252,28 @@ export class FrToasterContentComponent implements OnInit {
 
   public emitAction(): void {
     this.closed = true;
-    this._context.next();
+    this.toasterState = 'void';
+    this._context.next(<any>this.action as T);
+    if (this.onActionObserver) {
+      this.onActionObserver.next(this._context.action);
+      this.onActionObserver.complete();
+    }
   }
 
   ngOnInit() {
     this.toasterState = 'active';
+  }
 
-    // This is for animation
-    timer(this.timeout - 300).subscribe(() => {
-      if (!this.closed) {
-        this.toasterState = 'void';
-      }
-    });
-    // Close toaster after `timeout` milliseconds
-    timer(this.timeout).subscribe(() => {
-      if (!this.closed) {
-        this._context.complete();
-      }
+  ngOnDestroy() {
+    this._context.complete();
+    if (this.onActionObserver) {
+      this.onActionObserver.complete();
+    }
+  }
+
+  public onAction<T>(): Observable<T> {
+    return new Observable<T>((observer: Observer<T>) => {
+      this.onActionObserver = observer;
     });
   }
 }