+++ src/app/tooltip/index.ts
@@ -1,0 +1,2 @@
+export { FrTooltipDirective } from './tooltip.directive';
+export { FrTooltipModule } from './tooltip.module';